fdsdsg

+++ Beecrowd_Problems/Snack_1038.c
@@ -8,24 +8,24 @@ int main(){
     float x= (float)amount;
     if(code == 1){
 
-        int cachoro = 4.00*x;
-        printf("Total: R$ %d\n", cachoro);
+        float cachoro = 4.00*x;
+        printf("Total: R$ %.2f\n", cachoro);
     }else if(code == 2){
 
-        int salada = 4.50*x;
-        printf("Total: R$ %d\n", salada);
+        float salada = 4.50*x;
+        printf("Total: R$ %.2f\n", salada);
     }else if(code == 3){
 
-        int bacon = 5.00*x;
-        printf("Total: R$ %d\n", bacon);
+        float bacon = 5.00*x;
+        printf("Total: R$ %.2f\n", bacon);
     }else if(code == 4){
 
-        int torrada = 2.00*x;
-        printf("Total: R$ %d\n", torrada);
+        float torrada = 2.00*x;
+        printf("Total: R$ %.2f\n", torrada);
     }else if(code == 5){
 
-        int refriger = 1.50*x;
-        printf("Total: R$ %d\n", refriger);
+        float refriger = 1.50*x;
+        printf("Total: R$ %.2f\n", refriger);
     }
 
     return 0;
